Refactor: move bot_controller into controllers/main
